requirements.txt

responses
flask
flask-login
requests
python-dotenv
pytest
pytest-cov
flake8
